small http refactor added campaign
